requirements.txt
cement==3.0.6
jinja2
pyyaml
colorlog
python-amazon-sp-api
